fix scroll view
