+++ App/Components/ChannelDetail/index.js
@@ -1,6 +1,6 @@
 // @flow
 import React from 'react';
-import {Text, View, Image} from 'react-native';
+import {Text, View, Image, ScrollView, Dimensions} from 'react-native';
 import Icon from 'react-native-vector-icons/MaterialIcons';
 
 import styles from './style';
@@ -28,6 +28,13 @@ const InfoDetailItem = ({iconName, text}: TInfoDetailItem) => {
   );
 };
 
+const onScroll = (event, contentHeight, channelVideosGetRequest) => {
+  const windowScroll = Dimensions.get('window').height + event.nativeEvent.contentOffset.y;
+  if (windowScroll > contentHeight) {
+    channelVideosGetRequest();
+  }
+};
+
 type TChannelPanel = {
   channel: TDefaultChannel,
   channelActions: TChannelActions,
@@ -36,37 +43,47 @@ type TChannelPanel = {
 
 export const ChannelDetail = ({channel, likesPostRequest, channelActions}: TChannelPanel) => (
   <View style={styles.container}>
-    <View style={styles.infoContainer}>
-      <View style={styles.infoTop}>
-        <Image
-          style={styles.infoTopImage}
-          source={channel.item.youtube.thumbnail ? {uri: `${channel.item.youtube.thumbnail}`} : Images.mainWeak}
-        />
-        <View style={styles.infoTopLikeWrapper}>
-          <ButtonDefault
-            styles={styles.infoTopLike}
-            text={`いいね ${channel.item.likeCount}`}
-            onPress={() => likesPostRequest(channel.item)}
-            disabled={channel.isFetchingMyInfo}
-            iconName={'favorite'}
+    <ScrollView
+      onScroll={e => onScroll(
+        e, channel.contentHeightVideos,
+        () => channelActions.channelVideosGetRequest(channel.item.youtube.id),
+      )}
+      scrollEventThrottle={200}
+      onContentSizeChange={(width, height) => channelActions.channelVideosSetContentHeight(height)}
+      style={styles.scrollView}
+    >
+      <View style={styles.infoContainer}>
+        <View style={styles.infoTop}>
+          <Image
+            style={styles.infoTopImage}
+            source={channel.item.youtube.thumbnail ? {uri: `${channel.item.youtube.thumbnail}`} : Images.mainWeak}
           />
+          <View style={styles.infoTopLikeWrapper}>
+            <ButtonDefault
+              styles={styles.infoTopLike}
+              text={`いいね ${channel.item.likeCount}`}
+              onPress={() => likesPostRequest(channel.item)}
+              disabled={channel.isFetchingMyInfo}
+              iconName={'favorite'}
+            />
+          </View>
+        </View>
+        <View style={styles.infoName}>
+          <Text style={styles.infoNameText}>{channel.item.youtube.name}</Text>
+        </View>
+        <View style={styles.infoDetail}>
+          <InfoDetailItem iconName="trending-up" text={`ランキング ${channel.item.rank}位`} />
+          <InfoDetailItem iconName="trending-up" text={`マイランキング ${channel.itemMyInfo.rank}位`} />
+          <InfoDetailItem iconName="favorite" text={`総いいね数 ${channel.item.likeCount}`} />
+          <InfoDetailItem iconName="favorite" text={`マイいいね数 ${channel.itemMyInfo.likeCount}`} />
+          <InfoDetailItem iconName="subscriptions" text={`チャンネル登録者数 ${channel.item.youtube.subscriberCount}人`} />
         </View>
       </View>
-      <View style={styles.infoName}>
-        <Text style={styles.infoNameText}>{channel.item.youtube.name}</Text>
-      </View>
-      <View style={styles.infoDetail}>
-        <InfoDetailItem iconName="trending-up" text={`ランキング ${channel.item.rank}位`} />
-        <InfoDetailItem iconName="trending-up" text={`マイランキング ${channel.itemMyInfo.rank}位`} />
-        <InfoDetailItem iconName="favorite" text={`総いいね数 ${channel.item.likeCount}`} />
-        <InfoDetailItem iconName="favorite" text={`マイいいね数 ${channel.itemMyInfo.likeCount}`} />
-        <InfoDetailItem iconName="subscriptions" text={`チャンネル登録者数 ${channel.item.youtube.subscriberCount}人`} />
-      </View>
-    </View>
-    <Videos
-      channel={channel}
-      setContentHeight={channelActions.channelVideosSetContentHeight}
-      channelVideosGetRequest={channelActions.channelVideosGetRequest}
-    />
+      <Videos
+        channel={channel}
+        setContentHeight={channelActions.channelVideosSetContentHeight}
+        channelVideosGetRequest={channelActions.channelVideosGetRequest}
+      />
+    </ScrollView>
   </View>
 );
